+++ e.VOC.us/Game/Helper.cs
@@ -4,7 +4,7 @@ namespace e.VOC.us.Game
 {
     public static class Helper
     {
-        public static Vector2D AngleToUnitVector(int angle)
+        public static Vector2D AngleToUnitVector(float angle)
         {
             return new Vector2D((float)Math.Cos(angle / (180 / Math.PI)),(float)Math.Sin(angle / (180 / Math.PI)));
         }
